+++ online-store/src/components/productPage.ts
@@ -1,4 +1,5 @@
-const productPage: string = 
+import dataProducts from './dataProducts';
+const productPageLayout: string = 
 `
 <div class="wrapper">
   <p class="path-product font">STORE ></p>
@@ -41,4 +42,48 @@ const productPage: string =
   </div>
 </div>`
 
-export default productPage;
+export default function productPage(i: number): void{
+  let main = <HTMLElement>document.querySelector('.main');
+  main.innerHTML = productPageLayout;
+
+  let title = <HTMLElement>document.querySelector(".product-title");
+  title.innerText = dataProducts[i].title;
+
+  let description = <HTMLElement>(
+    document.querySelector(".product-description")
+  );
+  description.innerText = dataProducts[i].description;
+
+  let rating = <HTMLElement>document.querySelector(".product-rating");
+  rating.innerText = String(dataProducts[i].rating);
+
+  let stock = <HTMLElement>document.querySelector(".product-stock");
+  stock.innerText = String(dataProducts[i].stock);
+
+  let brand = <HTMLElement>document.querySelector(".product-brand");
+  brand.innerText = dataProducts[i].brand;
+
+  let category = <HTMLElement>document.querySelector(".product-category");
+  category.innerText = dataProducts[i].category;
+
+  let discount = <HTMLElement>(
+    document.querySelector(".product-discountPercentage")
+  );
+  discount.innerText = String(dataProducts[i].discountPercentage);
+
+  let path = <HTMLElement>document.querySelector(".path-product");
+  path.innerText = `Store > ${dataProducts[i].category} > ${dataProducts[i].brand} > ${dataProducts[i].title}`;
+
+  let main_page = <HTMLElement>document.querySelector(".photos-main");
+  main_page.style.background = `url('${dataProducts[i].images[0]}') no-repeat`;
+  for (let j = 0; j < 3; j++) {
+    let k = <HTMLElement>document.getElementsByClassName("photo-other")[j];
+    k.style.backgroundImage = `url('${dataProducts[i].images[j]}')`;
+    k.style.backgroundSize = "cover";
+    k.addEventListener("click", () => {
+      main_page.style.backgroundImage = `url('${dataProducts[i].images[j]}')`;
+    });
+  }
+  let cost = <HTMLElement>document.querySelector(".cost-product");
+  cost.innerText = `€${dataProducts[i].price}`;
+}
